Created WordleSolver2 which guesses the next word based on a words frequency in literature. Moved sorted lists of words to WordleCore. Removed public labels from IWordleSolver interface.

+++ WordleSolver/Controllers/SolverController.cs
@@ -31,7 +31,7 @@ namespace WordleSolver.Controllers
             }
             timer.Stop();
 
-            double guessesPerWordle = _guessesMade / _wordles.Count();
+            double guessesPerWordle = (double)_guessesMade / _wordles.Count();
             return new SolverResult(_solver.SolverIdentifier, guessesPerWordle, timer.ElapsedMilliseconds);
         }
